# routes/api.php
<?php

use App\Http\Controllers\Api\V1\AuthController;
use App\Http\Controllers\Api\V1\MeatController;
use App\Http\Controllers\Api\V1\AvailabilityController;
use App\Http\Controllers\Api\V1\ReservationController;
use App\Http\Controllers\Api\V1\Admin\MeatController as AdminMeatController;
use App\Http\Controllers\Api\V1\Admin\MeatItemController;
use App\Http\Controllers\Api\V1\Admin\AvailableDateController;
use App\Http\Controllers\Api\V1\Admin\PickupSlotController;
use App\Http\Controllers\Api\V1\Admin\ReservationController as AdminReservationController;
use Illuminate\Support\Facades\Route;

// Public routes
Route::prefix('v1')->group(function () {
    // Auth (com rate limiting)
    Route::post('/auth/register', [AuthController::class, 'register'])->middleware('throttle:5,60');
    Route::post('/auth/login', [AuthController::class, 'login'])->middleware('throttle:10,60');

    // Public catalog
    Route::get('/meats', [MeatController::class, 'index']);
    Route::get('/meats/{id}', [MeatController::class, 'show']);
    Route::get('/meats/{id}/items', [MeatController::class, 'items']);

    // Availability
    Route::get('/availability/dates', [AvailabilityController::class, 'dates']);
    Route::get('/availability/dates/{date}/slots', [AvailabilityController::class, 'slots']);

    // Public admin contact
    Route::get('/admin/whatsapp', function () {
        try {
            $admin = \App\Models\User::where('role', 'admin')->first();
            
            if (!$admin) {
                \Log::warning('No admin user found');
                return response()->json([
                    'whatsapp' => null,
                ], 200);
            }
            
            $whatsapp = $admin->whatsapp ?? null;
            
            // Limpar o número (remover espaços e caracteres especiais se necessário)
            if ($whatsapp) {
                $whatsapp = trim($whatsapp);
                // Se estiver vazio após trim, retornar null
                if ($whatsapp === '') {
                    $whatsapp = null;
                }
            }
            
            \Log::info('WhatsApp number fetched', [
                'admin_id' => $admin->id,
                'whatsapp' => $whatsapp ? 'present' : 'null'
            ]);
            
            return response()->json([
                'whatsapp' => $whatsapp,
            ], 200);
        } catch (\Exception $e) {
            \Log::error('Error fetching WhatsApp number: ' . $e->getMessage());
            return response()->json([
                'whatsapp' => null,
            ], 200);
        }
    });

    // Protected routes
    Route::middleware('auth:sanctum')->group(function () {
        // Auth
        Route::post('/auth/logout', [AuthController::class, 'logout']);
        Route::get('/auth/me', [AuthController::class, 'me']);
        Route::put('/auth/profile', [AuthController::class, 'updateProfile']);

        // Customer reservations
        Route::get('/reservations/my', [ReservationController::class, 'my']);
        Route::post('/reservations', [ReservationController::class, 'store']);
        Route::get('/reservations/{reservation}', [ReservationController::class, 'show']);
        Route::post('/reservations/{reservation}/cancel', [ReservationController::class, 'cancel']);

        // Admin routes
        Route::prefix('admin')->middleware('admin')->group(function () {
            // Meats
            Route::apiResource('meats', AdminMeatController::class);

            // Meat Items
            Route::get('/meat-items', [MeatItemController::class, 'index']);
            Route::get('/meat-items/{meatItem}', [MeatItemController::class, 'show']);
            Route::post('/meat-items/bulk', [MeatItemController::class, 'bulk']);
            Route::patch('/meat-items/{meatItem}', [MeatItemController::class, 'update']);
            Route::delete('/meat-items/{meatItem}', [MeatItemController::class, 'destroy']);

            // Available Dates
            Route::apiResource('available-dates', AvailableDateController::class);

            // Pickup Slots
            Route::post('/pickup-slots', [PickupSlotController::class, 'store']);
            Route::patch('/pickup-slots/{pickupSlot}', [PickupSlotController::class, 'update']);
            Route::delete('/pickup-slots/{pickupSlot}', [PickupSlotController::class, 'destroy']);

            // Reservations
            Route::get('/reservations', [AdminReservationController::class, 'index']);
            Route::post('/reservations/{reservation}/fulfill', [AdminReservationController::class, 'fulfill']);
        });
    });
});